requirements.txt
torch>=1.9.0
torchvision>=0.10.0
torchaudio>=0.9.0
opencv-python>=4.5.3
mediapipe>=0.8.9
transformers>=4.11.0
datasets>=1.11.0
evaluate>=0.2.2
openai>=0.27.0
espnet>=0.10.3
espnet_model_zoo>=0.1.0
soundfile>=0.10.3
matplotlib>=3.4.3
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=0.24.2
tqdm>=4.62.0
requests>=2.26.0
pyttsx3>=2.90